Cross-node communication examples.
In our Registry code review with Jose, we discussed how the current Registry stores a PID.  When a caller wants to make a cross-node call, the Registry gives them a PID from another node to make the call against.  Jose pointed out that this will be a bottleneck eventually, and that we should spawn new processes in remote nodes instead.

These changes reflect a very simple, but working way to demonstrate this approach.

Our next steps will be to refactor the Registry code to manage the various node names, and then provide those node names to the outbound calls.  We'll need to remove the coupling of the Service code from the Contexts and develop an updated Services or Relay Layer to establish a decoupled way for clients to issue calls to Engine without having to know what node they are or what node they want to delegate the work to.